mcp-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orca-orchestrator-mcp"
version = "0.1.0"
description = "Orca Network Orchestrator MCP Server"
authors = [
    {name = "Orca Network", email = "support@orca.network"}
]

requires-python = ">=3.10"
dependencies = [
    "mcp>=1.15.0",
    "supabase>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
orca-mcp-server = "orca_server.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"